+++ apps/echo-speaks.groovy
@@ -16,13 +16,13 @@
  */
 
 import groovy.transform.Field
-@Field static final String appVersionFLD  = "4.0.2.0"
-@Field static final String appModifiedFLD = "2021-01-27"
+@Field static final String appVersionFLD  = "4.0.3.0"
+@Field static final String appModifiedFLD = "2021-01-29"
 @Field static final String branchFLD      = "master"
 @Field static final String platformFLD    = "Hubitat"
 @Field static final Boolean betaFLD       = true
-@Field static final Boolean devModeFLD    = true
-@Field static final Map minVersionsFLD    = [echoDevice: 4020, wsDevice: 4020, actionApp: 4020, zoneApp: 4020, server: 270]  //These values define the minimum versions of code this app will work with.
+@Field static final Boolean devModeFLD    = false
+@Field static final Map minVersionsFLD    = [echoDevice: 4030, wsDevice: 4030, actionApp: 4030, zoneApp: 4030, server: 270]  //These values define the minimum versions of code this app will work with.
 
 @Field static final String sNULL          = (String)null
 @Field static final String sBLANK         = ''
@@ -34,6 +34,7 @@ import groovy.transform.Field
 @Field static final String sTRUE          = 'true'
 @Field static final String sBOOL          = 'bool'
 @Field static final String sENUM          = 'enum'
+@Field static final String sTIME          = 'time'
 @Field static final String sAPPJSON       = 'application/json'
 @Field static final String sIN_IGNORE     = 'In Ignore Device Input'
 @Field static final String sARM_AWAY      = 'ARMED_AWAY'
@@ -177,6 +178,7 @@ def mainPage() {
                         href "devCleanupPage", title: inTS("Removable Devices:"), description: "<div style='color: red; font-size: small;'>${rd}</div>", required: true, state: sNULL
                     }
                     String devDesc = getDeviceList()?.collect { """<span>${it?.value?.name}</span>${it?.value?.online ? """<span style='color: green;'> (Online)</span>""" : sBLANK}${it?.value?.supported == false ? """<span style='color: red;'> ${sFRNFACE}</span>""" : sBLANK}""" }?.sort().join("<br>").toString()
+/* """ */
                     String dd = devDesc ? """<div style='font-size: small; color: #1A77C9;'>${devDesc}</div><div style='font-weight: bold;font-size: small;'><br>${sTTM}</div>""" : """<div style='font-weight: bold; font-size: small;'>${sTTC}</div"""
                     href "deviceManagePage", title: inTS1("Manage Devices:", sDEVICES), description: dd, state: sCOMPLT
                 } else { paragraph pTS("Device Management will be displayed after install is complete", sNull, true, "orange") }
@@ -331,7 +333,7 @@ def servPrefPage() {
             input "resetService", sBOOL, title: inTS1("Reset Service Data?", sRESET), description: "This will clear all references to the current server and allow you to redeploy a new instance.\nLeave the page and come back after toggling.",
                 required: false, defaultValue: false, submitOnChange: true
             paragraph pTS("This will clear all references to the current server and allow you to redeploy a new instance.\nLeave the page and come back after toggling.", sNULL, false, sCLRGRY)
-            if(settings?.resetService) { clearCloudConfig() }
+            if(settings.resetService) { clearCloudConfig() }
         }
 /*        section(sTS("Documentation & Settings:")) {
             href url: documentationLink(), style: sEXTNRL, required: false, title: inTS1("View Documentation", "documentation"), description: sTTP
@@ -394,10 +396,10 @@ def alexaGuardPage() {
         section(sTS("Alexa Guard Control")) {
             input "alexaGuardAwayToggle", sBOOL, title: inTS1(gStateTitle, gStateIcon), description: "Current Status: ${gState}", defaultValue: false, submitOnChange: true
         }
-        if(settings?.alexaGuardAwayToggle != state.alexaGuardAwayToggle) {
+        if(settings.alexaGuardAwayToggle != state.alexaGuardAwayToggle) {
             setGuardState(settings.alexaGuardAwayToggle == true ? sARM_AWAY : sARM_STAY)
         }
-        state.alexaGuardAwayToggle = settings?.alexaGuardAwayToggle
+        state.alexaGuardAwayToggle = settings.alexaGuardAwayToggle
         section(sTS("Automate Guard Control")) {
             String t0 = guardAutoDesc()
             href "alexaGuardAutoPage", title: inTS1("Automate Guard Changes", "alarm_disarm"), description: t0, state: (t0==sTTC ? sNULL : sCOMPLT)
@@ -411,7 +413,7 @@ def alexaGuardAutoPage() {
         List amo = getAlarmModes()
         Boolean alarmReq = (settings.guardAwayAlarm || settings.guardHomeAlarm)
         Boolean modeReq = (settings.guardAwayModes || settings.guardHomeModes)
-        // Boolean swReq = (settings?.guardAwaySw || settings?.guardHomeSw)
+        // Boolean swReq = (settings.guardAwaySw || settings.guardHomeSw)
         section(sTS("Set Guard Using ${asn}")) {
             input "guardHomeAlarm", sENUM, title: inTS1("Home in ${asn} modes.", "alarm_home"), description: sTTS, options: amo, required: alarmReq, multiple: true, submitOnChange: true
             input "guardAwayAlarm", sENUM, title: inTS1("Away in ${asn} modes.", "alarm_away"), description: sTTS, options: amo, required: alarmReq, multiple: true, submitOnChange: true
@@ -503,7 +505,7 @@ def guardTriggerEvtHandler(evt) {
             setGuardHome()
         }
         if(newState == sARM_AWAY) {
-            if(settings?.guardAwayDelay) { logWarn("Setting Alexa Guard Mode to Away in (${settings?.guardAwayDelay} seconds)", true); runIn(settings?.guardAwayDelay, "setGuardAway") }
+            if(settings.guardAwayDelay) { logWarn("Setting Alexa Guard Mode to Away in (${settings.guardAwayDelay} seconds)", true); runIn(settings.guardAwayDelay, "setGuardAway") }
             else { setGuardAway(); logWarn("Setting Alexa Guard Mode to Away...", true) }
         }
     }
@@ -664,17 +666,16 @@ def zonesPage() {
         }
         section (sTS("Zone Management:"), hideable: true, hidden: true) {
             if(activeZones?.size()) {
-                input "pauseChildZones", sBOOL, title: inTS1("Pause all Zones?", "pause_orange"), description: "When pausing all Zones you can either restore all or open each zones and manually unpause it.",
-                        defaultValue: false, submitOnChange: true
+                input "pauseChildZones", sBOOL, title: inTS1("Pause all Zones?", "pause_orange"), description: "When pausing all Zones you can either restore all or open each zones and manually unpause it.", defaultValue: false, submitOnChange: true
                 if(settings.pauseChildZones) { settingUpdate("pauseChildZones", sFALSE, sBOOL); runIn(3, "executeZonePause") }
                 paragraph pTS("When pausing all zones you can either restore all or open each zone and manually unpause it.", sNULL, false, sCLRGRY)
             }
             if(pausedZones?.size()) {
                 input "unpauseChildZone", sBOOL, title: inTS1("Restore all actions?", "pause_orange"), defaultValue: false, submitOnChange: true
-                if(settings?.unpauseChildZones) { settingUpdate("unpauseChildZones", sFALSE, sBOOL); runIn(3, "executeZoneUnpause") }
+                if(settings.unpauseChildZone) { settingUpdate("unpauseChildZone", sFALSE, sBOOL); runIn(3, "executeZoneUnpause") }
             }
             input "reinitChildZones", sBOOL, title: inTS1("Clear Zones Status and force a full status refresh for all zones?", sRESET), defaultValue: false, submitOnChange: true
-            if(settings?.reinitChildZones) { settingUpdate("reinitChildZones", sFALSE, sBOOL); runIn(3, "executeZoneUpdate") }
+            if(settings.reinitChildZones) { settingUpdate("reinitChildZones", sFALSE, sBOOL); runIn(3, "executeZoneUpdate") }
         }
         state.childInstallOkFlag = true
         state.zoneDuplicated = false
@@ -989,11 +990,11 @@ def setNotificationTimePage() {
         section() {
             input "qStartInput", sENUM, title: inTS1("Starting at", "start_time"), options: ["A specific time", "Sunrise", "Sunset"], defaultValue: null, submitOnChange: true, required: false
             if(settings["qStartInput"] == "A specific time") {
-                input "qStartTime", "time", title: inTS1("Start time", "start_time"), required: timeReq
+                input "qStartTime", sTIME, title: inTS1("Start time", "start_time"), required: timeReq
             }
             input "qStopInput", sENUM, title: inTS1("Stopping at", "stop_time"), options: ["A specific time", "Sunrise", "Sunset"], defaultValue: null, submitOnChange: true, required: false
             if(settings?."qStopInput" == "A specific time") {
-                input "qStopTime", "time", title: inTS1("Stop time", "stop_time"), required: timeReq
+                input "qStopTime", sTIME, title: inTS1("Stop time", "stop_time"), required: timeReq
             }
             input "quietDays", sENUM, title: inTS1("Only on these week days", "day_calendar"), multiple: true, required: false, options: weekDaysEnum()
             input "quietModes", "mode", title: inTS1("When these modes are Active", "mode"), multiple: true, submitOnChange: true, required: false
@@ -1347,7 +1348,7 @@ def updated() {
 def initialize() {
     logInfo("running initialize...")
     //if(app?.getLabel() != "Echo Speaks") { app?.updateLabel("Echo Speaks") }
-    if((Boolean)settings.optOutMetrics && (String)state.appGuid) { if(removeInstallData()) { state.appGuid = sNULL } }
+    if((Boolean)settings.optOutMetrics && (String)state.appGuid) { if(removeInstallData()) { state.appGuid = sNULL; state.remove('appGuid') } }
     subscribe(location, "systemStart", startHandler)
     if(guardAutoConfigured()) {
         if(settings.guardAwayAlarm && settings.guardHomeAlarm) {
@@ -1424,10 +1425,10 @@ void appCleanup() {
     List items = [
         "availableDevices", "consecutiveCmdCnt", "isRateLimiting", "versionData", "heartbeatScheduled", "serviceAuthenticated", "cookie", "misPollNotifyWaitVal", "misPollNotifyMsgWaitVal",
         "updNotifyWaitVal", "lastDevActivity", "devSupMap", "tempDevSupData", "devTypeIgnoreData",
-        "warnHistory", "errorHistory", "bluetoothData", "dndData", "zoneStatusMap"
+        "warnHistory", "errorHistory", "bluetoothData", "dndData", "zoneStatusMap", "guardData", "guardDataSrc", "guardDataOverMaxSize"
     ]
-// pushTested
-    items?.each { String si-> if(state.containsKey(si)) { state.remove(si)} }
+    items.each { String si-> if(state.containsKey(si)) { state.remove(si)} }
+
     state.pollBlocked = false
     state.resumeConfig = false
     state.missPollRepair = false
@@ -1435,13 +1436,13 @@ void appCleanup() {
 
     // Settings Cleanup
     List setItems = ["performBroadcast", "stHub", "cookieRefreshDays"]
-    settings?.each { si-> ["music", "tunein", "announce", "perform", "broadcast", "sequence", "speech", "test_"]?.each { swi-> if(si?.key?.startsWith(swi as String)) { setItems?.push(si?.key as String) } } }
-    setItems?.unique()?.sort()?.each { sI-> if(settings?.containsKey(sI as String)) { settingRemove(sI as String) } }
+    settings?.each { si-> ["music", "tunein", "announce", "perform", "broadcast", "sequence", "speech", "test_"].each { String swi-> if(si.key?.startsWith(swi)) { setItems.push(si?.key as String) } } }
+    setItems.unique().sort().each { String sI-> if(settings?.containsKey(sI)) { settingRemove(sI) } }
     cleanUpdVerMap()
 }
 
 void wsEvtHandler(evt) {
-    // log.trace("wsEvtHandler evt: ${evt}")
+    if(devModeFLD) logTrace("wsEvtHandler evt: ${evt}")
     if(evt && evt.id && (evt.attributes?.size() || evt.triggers?.size())) {
         if("bluetooth" in evt.triggers) { runIn(2, "getBluetoothRunIn") } // getBluetoothDevices(true)
         if("activity" in evt.triggers) { runIn(1, "getDeviceActivityRunIn") } // Map a=getDeviceActivity(sNULL, true)
@@ -1903,6 +1904,7 @@ void authEvtHandler(Boolean isAuth, String src=sNULL) {
     Boolean stC = ((Boolean)state.authValid != isAuth)
     logDebug "authEvtHandler(${isAuth},$src) stateChange: ${stC}"
     state.authValid = isAuth
+    state.remove('authEvtClearReason')
     if(!isAuth && !(Boolean)state.noAuthActive) {
         state.noAuthActive = true
         clearCookieData('authHandler', true)
@@ -2546,7 +2548,7 @@ void checkGuardSupportResponse(response, data) {
             } else {
                 logWarn("Can't check for Guard Support because server version is out of date...  Please update to the latest version...")
             }
-            state.guardDataOverMaxSize = true
+            state.alexaGuardDataOverMaxSize = true
             return
         }
         Map resp = response?.data ? parseJson(response?.data?.toString()) : null
@@ -2555,10 +2557,14 @@ void checkGuardSupportResponse(response, data) {
             Map locDetails = details?.locationDetails?.locationDetails?.Default_Location?.amazonBridgeDetails?.amazonBridgeDetails["LambdaBridge_AAA/OnGuardSmartHomeBridgeService"] ?: null
             if(locDetails && locDetails?.applianceDetails && locDetails?.applianceDetails?.applianceDetails) {
                 def guardKey = locDetails?.applianceDetails?.applianceDetails?.find { it?.key?.startsWith("AAA_OnGuardSmartHomeBridgeService_") }
+// could there be multiple Guards?
+                def guardKeys = locDetails?.applianceDetails?.applianceDetails?.findAll { it?.key?.startsWith("AAA_OnGuardSmartHomeBridgeService_") }
+                if(devModeFLD) logTrace("Guardkeys: ${guardKeys.size()}")
                 def guardData = locDetails?.applianceDetails?.applianceDetails[guardKey?.key]
-                // log.debug "Guard: ${guardData}"
+                if(devModeFLD) logTrace("Guard: ${guardData}")
                 if(guardData?.modelName == "REDROCK_GUARD_PANEL") {
-                    state.guardData = [
+//TODO: we really need to match guardData to devices (and really locations)  ie guard can be on some devices/locations and not on others
+                    state.alexaGuardData = [
                         entityId: guardData?.entityId,
                         applianceId: guardData?.applianceId,
                         friendlyName: guardData?.friendlyName,
@@ -2572,7 +2578,7 @@ void checkGuardSupportResponse(response, data) {
     }
     state.alexaGuardSupported = guardSupported
     updTsVal("lastGuardSupChkDt")
-    state.guardDataSrc = "app"
+    state.alexaGuardDataSrc = "app"
     if(guardSupported) getGuardState()
 }
 
@@ -2601,7 +2607,7 @@ void checkGuardSupportServerResponse(response, data) {
             if(resp) {
                 if(resp.guardData) {
                     log.debug "AGS Server Resp: ${resp?.guardData}"
-                    state.guardData = resp.guardData
+                    state.alexaGuardData = resp.guardData
                     guardSupported = true
                 }
             } else { logError("checkGuardSupportServerResponse Error | No data received..."); return }
@@ -2611,8 +2617,8 @@ void checkGuardSupportServerResponse(response, data) {
         return
     }
     state.alexaGuardSupported = guardSupported
-    state.guardDataOverMaxSize = guardSupported
-    state.guardDataSrc = "server"
+    state.alexaGuardDataOverMaxSize = guardSupported
+    state.alexaGuardDataSrc = "server"
     updTsVal("lastGuardSupChkDt")
     if(guardSupported) getGuardState()
 }
@@ -2626,7 +2632,7 @@ void getGuardState() {
         headers: getCookieMap(),
         contentType: sAPPJSON,
         timeout: 20,
-        body: [ stateRequests: [ [entityId: state.guardData?.applianceId, entityType: "APPLIANCE" ] ] ]
+        body: [ stateRequests: [ [entityId: state.alexaGuardData?.applianceId, entityType: "APPLIANCE" ] ] ]
     ]
     try {
         logTrace("getGuardState")
@@ -2634,8 +2640,12 @@ void getGuardState() {
             if(resp?.status != 200) logWarn("${resp?.status} $params")
             if(resp?.status == 200) updTsVal("lastSpokeToAmazon")
             Map respData = resp?.data ?: null
+
+            if(devModeFLD) log.debug "GuardState resp: ${respData}"
+
             if(respData && respData?.deviceStates && respData?.deviceStates[0] && respData?.deviceStates[0]?.capabilityStates) {
                 def guardStateData = parseJson(respData?.deviceStates[0]?.capabilityStates as String)
+                if(devModeFLD) logTrace("guardState: ${guardStateData}")
                 String curState = (String)state.alexaGuardState ?: sNULL
                 state.alexaGuardState = guardStateData?.value[0] ? (String)guardStateData?.value[0] : (String)guardStateData?.value
                 settingUpdate("alexaGuardAwayToggle", (((String)state.alexaGuardState == sARM_AWAY) ? sTRUE : sFALSE), sBOOL)
@@ -2643,7 +2653,6 @@ void getGuardState() {
                 if(curState != (String)state.alexaGuardState) updGuardActionTrig()
                 updTsVal("lastGuardStateChkDt")
             }
-            // log.debug "GuardState resp: ${respData}"
         }
     } catch (ex) {
         respExceptionHandler(ex, "getGuardState", true)
@@ -2657,7 +2666,7 @@ void setGuardState(String guardState) {
     guardState = guardStateConv(guardState)
     logDebug("setAlexaGuard($guardState)")
     try {
-        String body = new groovy.json.JsonOutput()?.toJson([ controlRequests: [ [ entityId: state.guardData?.applianceId as String, entityType: "APPLIANCE", parameters: [action: "controlSecurityPanel", armState: guardState ] ] ] ])
+        String body = new groovy.json.JsonOutput()?.toJson([ controlRequests: [ [ entityId: state.alexaGuardData?.applianceId as String, entityType: "APPLIANCE", parameters: [action: "controlSecurityPanel", armState: guardState ] ] ] ])
         Map params = [
             uri: getAmazonUrl(),
             path: "/api/phoenix/state",
@@ -2923,7 +2932,7 @@ void receiveEventData(Map evtData, String src) {
                 List unknownDevices = []
                 List curDevFamily = []
 //                Integer cnt = 0
-                String devAcctId = (String)null
+                String devAcctId = sNULL
                 evtData?.echoDevices?.each { String echoKey, echoValue->
                     devAcctId = echoValue?.deviceAccountId
                     logTrace("echoDevice | $echoKey | ${echoValue}")
@@ -2973,7 +2982,7 @@ void receiveEventData(Map evtData, String src) {
                     echoValue["amazonDomain"] = (settings.amazonDomain ?: "amazon.com")
                     echoValue["regionLocale"] = (settings.regionLocale ?: "en-US")
                     echoValue["cookie"] = getCookieMap()
-                    echoValue["deviceAccountId"] = echoValue?.deviceAccountId as String ?: (String)null
+                    echoValue["deviceAccountId"] = echoValue?.deviceAccountId as String ?: sNULL
                     echoValue["deviceStyle"] = deviceStyleData
                     // log.debug "deviceStyle: ${echoValue?.deviceStyle}"
 
@@ -4058,7 +4067,7 @@ private getDiagDataJson(Boolean asObj = false) {
                 cookieValidHistory: (List)state.authValidHistory,
                 cookieLastRefreshDate: getTsVal("lastCookieRrshDt") ?: null,
                 cookieLastRefreshDur: getTsVal("lastCookieRrshDt") ? seconds2Duration(getLastTsValSecs("lastCookieRrshDt")) : null,
-                cookieInvalidReason: (!(Boolean)state.authValid && state.authEvtClearReason) ? state.authEvtClearReason : (String)null,
+                cookieInvalidReason: (!(Boolean)state.authValid && state.authEvtClearReason) ? state.authEvtClearReason : sNULL,
                 cookieRefreshDays: (Integer)settings.refreshCookieDays,
                 cookieItems: [
                     hasLocalCookie: (state.cookieData && state.cookieData.localCookie),
@@ -4077,11 +4086,11 @@ private getDiagDataJson(Boolean asObj = false) {
             alexaGuard: [
                 supported: (Boolean)state.alexaGuardSupported,
                 status: (String)state.alexaGuardState,
-                dataSrc: (String)state.guardDataSrc,
+                dataSrc: (String)state.alexaGuardDataSrc,
                 lastSupportCheck: getTsVal("lastGuardSupChkDt"),
                 lastStateCheck: getTsVal("lastGuardStateChkDt"),
                 lastStateUpd: getTsVal("lastGuardStateUpdDt"),
-                stRespLimit: (Boolean)state.guardDataOverMaxSize
+                stRespLimit: (Boolean)state.alexaGuardDataOverMaxSize
             ],
             server: [
                 version: state.codeVersions?.server ?: null,
@@ -5577,7 +5586,7 @@ void logError(String msg, Boolean noHist=false, ex=null) {
     if(!noHist) { addToLogHistory("errorHistory", msg, 15) }
 }
 
-String addHead(String msg) {
+static String addHead(String msg) {
     return "EchoApp (v"+appVersionFLD+") | "+msg
 }