+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## _**(January 29th, 2021)**_
+### _**Echo Speaks App (v4.0.3.0)**_
+- [FIX] bug fixes.
+
+### _**Echo Speaks Actions (v4.0.3.0)**_
+- [FIX] bug fixes.
+
+### _**Echo Speaks Zones (v4.0.3.0)**_
+- [FIX] bug fixes.
+
+### _**Echo Speaks Device (v4.0.3.0)**_
+- [FIX] bug fixes.
+
+### _**Echo Speaks WebSocket (v4.0.3.0)**_
+- [FIX] bug fixes.
+
 ## _**(January 27th, 2021)**_
 ### _**Echo Speaks App (v4.0.2.0)**_
 - [UPDATE] Slowly modifying the app UI to use HTML.